// app/src/main/webui/src/api/model/problemWorkflowDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Homelab Manager API
 * OpenAPI spec version: 0.0.1
 */
import type { Uuid } from "./uuid";
import type { ProblemWorkflowDtoProblemType } from "./problemWorkflowDtoProblemType";
import type { ProblemWorkflowDtoDescription } from "./problemWorkflowDtoDescription";
import type { ProblemWorkflowDtoMedia } from "./problemWorkflowDtoMedia";
import type { ProblemWorkflowDtoGrabbedRelease } from "./problemWorkflowDtoGrabbedRelease";
import type { LocalDateTime } from "./localDateTime";
import type { ProblemWorkflowDtoCompletedAt } from "./problemWorkflowDtoCompletedAt";

export interface ProblemWorkflowDto {
  id: Uuid;
  mediaType: string;
  status: string;
  currentStep: string;
  problemType: ProblemWorkflowDtoProblemType;
  description: ProblemWorkflowDtoDescription;
  media: ProblemWorkflowDtoMedia;
  grabbedRelease: ProblemWorkflowDtoGrabbedRelease;
  createdAt: LocalDateTime;
  updatedAt: LocalDateTime;
  completedAt: ProblemWorkflowDtoCompletedAt;
}